Move error related functions to errors.ts
