+++ asset-transfer-basic/rest-api-typescript/src/errors.ts
@@ -2,6 +2,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import { logger } from './logger';
+
 export class TransactionError extends Error {
   transactionId: string;
 
@@ -43,3 +45,91 @@ export class AssetNotFoundError extends TransactionError {
     this.name = 'AssetNotFoundError';
   }
 }
+
+/*
+ * Checks whether an error was caused by a duplicate transaction.
+ *
+ * Checking error strings like this is not ideal, unfortunately it appears to
+ * be the only option. In this case it would be better to check for the
+ * DUPLICATE_TXID TxValidationCode somehow but that does not seem to be
+ * possible.
+ */
+export const isDuplicateTransactionError = (error: {
+  errors: { endorsements: { details: string }[] }[];
+}): boolean => {
+  try {
+    const isDuplicateTxn = error?.errors?.some((err) =>
+      err?.endorsements?.some((endorsement) =>
+        endorsement?.details?.startsWith('duplicate transaction found')
+      )
+    );
+
+    return isDuplicateTxn;
+  } catch (err) {
+    logger.warn(err, 'Error checking for duplicate transaction');
+  }
+
+  return false;
+};
+
+/*
+ * Handles errors from evaluating and submitting transactions.
+ *
+ * As with duplicate transaction errors, checking error strings like this is
+ * not ideal. Unfortunately the chaincode samples do not use error codes so
+ * again it's the only option. The error message text is not even the same for
+ * the Go, Java, and Javascript implementations of the chaincode!
+ */
+export const handleError = (transactionId: string, err: Error): Error => {
+  // This regex needs to match the following error messages:
+  //   "the asset %s already exists"
+  //   "The asset ${id} already exists"
+  //   "Asset %s already exists"
+  const assetAlreadyExistsRegex = /([tT]he )?[aA]sset \w* already exists/g;
+  const assetAlreadyExistsMatch = err.message.match(assetAlreadyExistsRegex);
+  logger.debug(
+    { message: err.message, result: assetAlreadyExistsMatch },
+    'Checking for asset already exists message'
+  );
+  if (assetAlreadyExistsMatch) {
+    return new AssetExistsError(assetAlreadyExistsMatch[0], transactionId);
+  }
+
+  // This regex needs to match the following error messages:
+  //   "the asset %s does not exist"
+  //   "The asset ${id} does not exist"
+  //   "Asset %s does not exist"
+  const assetDoesNotExistRegex = /([tT]he )?[aA]sset \w* does not exist/g;
+  const assetDoesNotExistMatch = err.message.match(assetDoesNotExistRegex);
+  logger.debug(
+    { message: err.message, result: assetDoesNotExistMatch },
+    'Checking for asset does not exist message'
+  );
+  if (assetDoesNotExistMatch) {
+    return new AssetNotFoundError(assetDoesNotExistMatch[0], transactionId);
+  }
+
+  // This regex needs to match the following error messages:
+  //   "Failed to get transaction with id %s, error Entry not found in index"
+  const transactionDoesNotExistRegex =
+    /Failed to get transaction with id [^,]*, error Entry not found in index/g;
+  const transactionDoesNotExistMatch = err.message.match(
+    transactionDoesNotExistRegex
+  );
+  logger.debug(
+    { message: err.message, result: transactionDoesNotExistMatch },
+    'Checking for transaction does not exist message'
+  );
+  if (transactionDoesNotExistMatch) {
+    return new TransactionNotFoundError(
+      transactionDoesNotExistMatch[0],
+      transactionId
+    );
+  }
+
+  logger.error(
+    { transactionId: transactionId, error: err },
+    'Unhandled transaction error'
+  );
+  return new TransactionError('Transaction error', transactionId);
+};
