+++ asset-transfer-basic/rest-api-typescript/src/fabric.ts
@@ -25,12 +25,7 @@ import {
   incrementRetryCount,
   TransactionDetails,
 } from './redis';
-import {
-  AssetExistsError,
-  AssetNotFoundError,
-  TransactionError,
-  TransactionNotFoundError,
-} from './errors';
+import { handleError, isDuplicateTransactionError } from './errors';
 import protos from 'fabric-protos';
 
 /*
@@ -247,63 +242,6 @@ export const submitTransaction = async (
   return txnId;
 };
 
-// Unfortunately the chaincode samples do not use error codes, and the error
-// message text is not the same for each implementation
-// TODO move to errors.ts?
-const handleError = (transactionId: string, err: Error): Error => {
-  // This regex needs to match the following error messages:
-  //   "the asset %s already exists"
-  //   "The asset ${id} already exists"
-  //   "Asset %s already exists"
-  const assetAlreadyExistsRegex = /([tT]he )?[aA]sset \w* already exists/g;
-  const assetAlreadyExistsMatch = err.message.match(assetAlreadyExistsRegex);
-  logger.debug(
-    { message: err.message, result: assetAlreadyExistsMatch },
-    'Checking for asset already exists message'
-  );
-  if (assetAlreadyExistsMatch) {
-    return new AssetExistsError(assetAlreadyExistsMatch[0], transactionId);
-  }
-
-  // This regex needs to match the following error messages:
-  //   "the asset %s does not exist"
-  //   "The asset ${id} does not exist"
-  //   "Asset %s does not exist"
-  const assetDoesNotExistRegex = /([tT]he )?[aA]sset \w* does not exist/g;
-  const assetDoesNotExistMatch = err.message.match(assetDoesNotExistRegex);
-  logger.debug(
-    { message: err.message, result: assetDoesNotExistMatch },
-    'Checking for asset does not exist message'
-  );
-  if (assetDoesNotExistMatch) {
-    return new AssetNotFoundError(assetDoesNotExistMatch[0], transactionId);
-  }
-
-  // This regex needs to match the following error messages:
-  //   "Failed to get transaction with id %s, error Entry not found in index"
-  const transactionDoesNotExistRegex =
-    /Failed to get transaction with id [^,]*, error Entry not found in index/g;
-  const transactionDoesNotExistMatch = err.message.match(
-    transactionDoesNotExistRegex
-  );
-  logger.debug(
-    { message: err.message, result: transactionDoesNotExistMatch },
-    'Checking for transaction does not exist message'
-  );
-  if (transactionDoesNotExistMatch) {
-    return new TransactionNotFoundError(
-      transactionDoesNotExistMatch[0],
-      transactionId
-    );
-  }
-
-  logger.error(
-    { transactionId: transactionId, error: err },
-    'Unhandled transaction error'
-  );
-  return new TransactionError('Transaction error', transactionId);
-};
-
 /*
  * Retry a transaction
  *
@@ -357,26 +295,6 @@ const retryTransaction = async (
   }
 };
 
-// TODO move to errors.ts?
-const isDuplicateTransactionError = (error: {
-  errors: { endorsements: { details: string }[] }[];
-}) => {
-  // TODO this is horrible! Isn't it possible to check for TxValidationCode DUPLICATE_TXID somehow?
-  try {
-    const isDuplicateTxn = error?.errors?.some((err) =>
-      err?.endorsements?.some((endorsement) =>
-        endorsement?.details?.startsWith('duplicate transaction found')
-      )
-    );
-
-    return isDuplicateTxn;
-  } catch (err) {
-    logger.warn(err, 'Error checking for duplicate transaction');
-  }
-
-  return false;
-};
-
 /*
  * Block event listener to handle successful transactions
  *
@@ -409,9 +327,9 @@ export const blockEventHandler = (redis: Redis): BlockListener => {
 
 /*
  * Get the current block height
- * 
+ *
  * This example of using a system contract is used for the liveness REST
- * endpoint 
+ * endpoint
  */
 export const getBlockHeight = async (
   qscc: Contract
